+++ main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-const version = "1.7.2"
+const version = "1.7.3"
 
 func main() {